// graphql-types.ts
export type Language =
  | 'EN'
  | 'RU';
// ...
export type SeoQueryQuery = { site?: { siteMetadata?: { i18n?: Array<{ lang?: Language | null, siteMetadata?: { title?: string | null, description?: string | null, author?: { name?: string | null, summary?: string | null } | null } | null } | null> | null } | null } | null };